# TypeTree speedup shiftindicies