--- a/enzyme/Enzyme/TypeAnalysis/TypeTree.h
+++ b/enzyme/Enzyme/TypeAnalysis/TypeTree.h
@@ -992,8 +992,7 @@ public:
     Result.minIndices.resize(maxInsertedDepth);
 
     auto Res2 = ShiftIndicesSlow(dl, offset, maxSize, addOffset);
-    assert(Result == Res2);
-    if (Result.minIndices != Res2.minIndices) {
+    if (Result.minIndices != Res2.minIndices || !(Result == Res2)) {
       llvm::errs() << " off: " << offset << " maxSize: " << maxSize
                    << " addOffset: " << addOffset << "\n";
       llvm::errs() << "inp: " << str() << "\n";
@@ -1003,6 +1002,7 @@ public:
                    << "\n";
       llvm::errs() << " res2 minidx : " << to_string(Res2.minIndices) << "\n";
     }
+    assert(Result == Res2);
     assert(Result.minIndices == Res2.minIndices);
 
     return Result;