Fixes https://github.com/EnzymeAD/Enzyme/issues/1743