@wsmoses 2024-02-19
note I left the slow version right now to do correctness tests, which already found a bug locally (which have fixed). Going to run it on the whole test suite, then disable the check once it comes back valid.