@wsmoses 2024-02-19
Also note this fix hopefully changes the asymptotic behavior of ShiftIndices from $O(n^2)$ to $O(n)$ [which it really should've been the whole time]