@vchuravy 2024-02-20
Seems like a speedup from 220s to 180s, but `TypeTree::insert` and `TypeTree:checkedOrIn` are still very hot

Before:

![image](https://github.com/EnzymeAD/Enzyme/assets/145258/1dfc562c-1d09-442d-8411-71d6cb927862)


After:

![image](https://github.com/EnzymeAD/Enzyme/assets/145258/444828d7-3373-4fa6-a88b-d30eed5f3d80)

![image](https://github.com/EnzymeAD/Enzyme/assets/145258/b6c1e6f6-dd2b-4a12-a304-41a28fa97267)

Detailed analysis:

- `checkedOrIn` ~28s (42% of compilation time)
    - 17s from update Analysis 
    - 10.5s from EnzymeMergeTypeTree
- `CanonicalizeInPlace`: 11s