@ZuseZ4 2024-02-20
@vchuravy If already at it, could you please also profile 
https://github.com/EnzymeAD/Enzyme/pull/1641 if you have a few minutes? My server never had the permissions for vtune. My new Laptop is reasonable otherwise, so I can also set it up myself