@wsmoses 2024-02-20
Considering I don’t see ShiftIndices more on the flame graph I’m going to
call this a win.

I’ll see what I can do about the Julia function. mergeTypeTree is the C API
and primarily used in that function so I think it’s the next most
meaningful place we should be able to reduce overhead from. Essentially
because it should be not terrible to make it called only once. CheckedOrIn
however doesn’t have any obvious speed ups I remember.

On Tue, Feb 20, 2024 at 2:02 AM Manuel Drehwald ***@***.***>
wrote:

> @vchuravy <https://github.com/vchuravy> If already at it, could you
> please also profile
> #1641 <https://github.com/EnzymeAD/Enzyme/pull/1641> if you have a few
> minutes? My server never had the permissions for vtune. My new Laptop is
> reasonable otherwise, so I can also set it up myself
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/EnzymeAD/Enzyme/pull/1744#issuecomment-1953595354>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAJTUXDS4ZNE23F3VAS4O43YURC7BAVCNFSM6AAAAABDQF76POVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTSNJTGU4TKMZVGQ>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>